implement inverse trans for Hittable, significant performance boost
Computing the inverse matrix each time is a huge performance
bottleneck of course. This adds a separate field that is updated with
the inverse transformation of `trans` each time `trans` is
modified (which only really happens during scene construction).

Also this marks the `Hittable` type as `acyclic` to make ORC happy. We
don't form cycles from `Hittables` so this is fine.