clean up telescope construction code
Note: fixes a small bug in the telescope placement. The old code had a
`-sin(α) lM/2` term there, which was too much! That displacement
already takes place in `setCone` as the - what is now called -
xOffset!

The result of the old code was that the lowest layer actually touched
the bottom part of the bore *at the front* (top part)