gigantic refactor to replace `Color` by `Spectrum`
`Spectrum` is a generic type with a statically known size. It can be
an RGB color or a generalized spectrum with `NumSamples` (adjustable
via `-d:NumSamples=<integer>`.

Excuse the gigantic commit, but thanks to how Nim generics work it's
tricky to make these changes more atomically.

The RenderContext knows different HittablesLists now, for the Camera
as well as for the axion emission.

This also adds ~XrayMatter~ as a Material, which contains
~AngleInterpolator~, which is used to store the correct spectra for
different incidence angle on an X-ray mirror.

Finally also adds different modes for the ~Sensor2D~ (as ImageSensor)
to work in. Either regular counting mode ~sCount~ or ~sSum~, which
sums up the incident spectra for each pixel.